[dbus-c++] fix spurious wakeups
I was seeing spurious debugd and chapsd wakeups every 10 seconds.
Root caused it to the use of wait_min to upper bound the poll
timeout to 10 seconds. There is really no good reason to do this.
Anyone using this eventloop just wakes up, checks all timeouts
and fds, finds there is nothing to do and then goes back to sleep.

BUG=chromium:227114
TEST=Verify debugd still works via chrome://system.
TEST=Verify that chapsd and debugd no longer generate spurious wakes.

Change-Id: I381dc26917bd334be1ac8fe394460af3be4de70a
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47951
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Elly Jones <ellyjones@chromium.org>