dbus-c++: remove overloading for async proxy methods
We'd like to take the addresses of asynchronous proxy
methods in shill, and overloading makes that difficult.

In particular, when the synchronous and asynchronous
methods share a name, we must case the name to the
asynchronous method's type, in order to get the
address of the asynchronous method.

Disambiguate the synchronous and asynchronous methods,
by appending "Async" to the name of the asynchronous
methods. This makes it possible to simply say
&<ProxyClass>::MethodAsync.

CQ-DEPEND=CL:195762
BUG=chromium:365390
TEST=emerge dbus-c++, platform2, and mtpd

Change-Id: Ida3176e37e6ba27a53df62d41afb23592784f188
Reviewed-on: https://chromium-review.googlesource.com/195761
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>