Report child nodes of the "/" path
The code in dbus-c++ which searches for child nodes of a path did
not special case the prefix used for searching for children of "/".
Add this code, as well as adding another special case to avoid
including "" as a child of "/" (since the "/" is a strict substring
of "/").

BUG=chromium:394470
TEST=gdbus introspect --system --dest org.chromium.flimflam --object-path / --recurse --only-properties
dbus-send --system --dest=org.chromium.flimflam --type=method_call --print-reply / org.freedesktop.DBus.Introspectable.Introspect

Change-Id: I622603bcbf6dcec2c1541f77fb10c60b2982ca81
Reviewed-on: https://chromium-review.googlesource.com/211617
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>