Handle exceptions in async mode.
Previously, setting the flag AVOID_EXCEPTIONS enabled the possbility to
handle service requests asynchronously. Unforturnately, it also removed the
handling of exceptions. Because of this, it was easy to crash the
application by sending bad D-Bus data to it.

Changed the name of the exception_flag to async_flag. Added two extra
options, ASYNCRONOUS and ASYNCRONOUS_AVOID_EXCEPTIONS, where the latter is
equivalent to the old AVOID_EXCEPTIONS flag. To maintain backward
compatibility, the old AVOID_EXCEPTIONS flag is still present.